27.12.18 22:00

+++ expressions/Expression.h
@@ -5,7 +5,7 @@
 
 class Expression{
 public:
-    virtual double calculate() =0;
+    virtual float calculate() =0;
     virtual ~Expression() = default;
 };
